Move Loading composable into "designsystem" package.
